printOrders method and maxSliceSize condition
Added a printOrders method to the OrderManager class to allow us to see which orders are in the 'orders' hashmap at any point.

Added a maximum slice size condition to the price method to replace the original method of simply halving the sizeRemaining. This max slice size can easily be changed, but allows us to have a maximum trade condition just like in the real markets (to avoid market volatility).