Update on Friday tasks
Have added a logInfo method to MyLogger, which prints a given string to the trades logFile. I have implemented this to add a starting line each time the program is run.

I have added an Order constructor which takes price into account. This in turn has allowed me to link the fill price to the initialMarketPrice, and I am assuming that the price displayed is per unit of the given instrument.

In SampleClient, I have implemented a hashmap constructor to store the available instruments and their initial market price based on the SampleLiveMarketData.

I have changed the current maxSliceSize from 100 to 1000.